Add tests for mac and tvos
Changes:
- Renames DemoTests folder to tests
- Add test targets for OS X and tvOS
- Update Podfile to have the new targets
- Update Rakefile to run all the new test targets
