+++ Demo/Demo.xcodeproj/xcshareddata/xcschemes/MoyaTests-Mac.xcscheme
@@ -9,28 +9,14 @@
          <BuildActionEntry
             buildForTesting = "YES"
             buildForRunning = "YES"
-            buildForProfiling = "YES"
-            buildForArchiving = "YES"
-            buildForAnalyzing = "YES">
-            <BuildableReference
-               BuildableIdentifier = "primary"
-               BlueprintIdentifier = "5EC197DF1A1BB16D00F4DFD4"
-               BuildableName = "Demo.app"
-               BlueprintName = "Demo"
-               ReferencedContainer = "container:Demo.xcodeproj">
-            </BuildableReference>
-         </BuildActionEntry>
-         <BuildActionEntry
-            buildForTesting = "YES"
-            buildForRunning = "NO"
             buildForProfiling = "NO"
             buildForArchiving = "NO"
             buildForAnalyzing = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "5EC197F61A1BB16D00F4DFD4"
-               BuildableName = "MoyaTests-iOS.xctest"
-               BlueprintName = "MoyaTests-iOS"
+               BlueprintIdentifier = "063D07EC1C0D362200C7061E"
+               BuildableName = "MoyaTests-Mac.xctest"
+               BlueprintName = "MoyaTests-Mac"
                ReferencedContainer = "container:Demo.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
@@ -47,9 +33,9 @@
             skipped = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "5EC197F61A1BB16D00F4DFD4"
-               BuildableName = "MoyaTests-iOS.xctest"
-               BlueprintName = "MoyaTests-iOS"
+               BlueprintIdentifier = "063D07EC1C0D362200C7061E"
+               BuildableName = "MoyaTests-Mac.xctest"
+               BlueprintName = "MoyaTests-Mac"
                ReferencedContainer = "container:Demo.xcodeproj">
             </BuildableReference>
          </TestableReference>
@@ -57,9 +43,9 @@
       <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "5EC197DF1A1BB16D00F4DFD4"
-            BuildableName = "Demo.app"
-            BlueprintName = "Demo"
+            BlueprintIdentifier = "063D07EC1C0D362200C7061E"
+            BuildableName = "MoyaTests-Mac.xctest"
+            BlueprintName = "MoyaTests-Mac"
             ReferencedContainer = "container:Demo.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
@@ -76,16 +62,15 @@
       debugDocumentVersioning = "YES"
       debugServiceExtension = "internal"
       allowLocationSimulation = "YES">
-      <BuildableProductRunnable
-         runnableDebuggingMode = "0">
+      <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "5EC197DF1A1BB16D00F4DFD4"
-            BuildableName = "Demo.app"
-            BlueprintName = "Demo"
+            BlueprintIdentifier = "063D07EC1C0D362200C7061E"
+            BuildableName = "MoyaTests-Mac.xctest"
+            BlueprintName = "MoyaTests-Mac"
             ReferencedContainer = "container:Demo.xcodeproj">
          </BuildableReference>
-      </BuildableProductRunnable>
+      </MacroExpansion>
       <AdditionalOptions>
       </AdditionalOptions>
    </LaunchAction>
@@ -95,16 +80,15 @@
       savedToolIdentifier = ""
       useCustomWorkingDirectory = "NO"
       debugDocumentVersioning = "YES">
-      <BuildableProductRunnable
-         runnableDebuggingMode = "0">
+      <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "5EC197DF1A1BB16D00F4DFD4"
-            BuildableName = "Demo.app"
-            BlueprintName = "Demo"
+            BlueprintIdentifier = "063D07EC1C0D362200C7061E"
+            BuildableName = "MoyaTests-Mac.xctest"
+            BlueprintName = "MoyaTests-Mac"
             ReferencedContainer = "container:Demo.xcodeproj">
          </BuildableReference>
-      </BuildableProductRunnable>
+      </MacroExpansion>
    </ProfileAction>
    <AnalyzeAction
       buildConfiguration = "Debug">
