+++ Demo/Tests/AlamofireMoyaMappingSpec.swift
