+++ Rakefile
@@ -1,27 +1,67 @@
-def xcodebuild_in_demo_dir(tasks, xcprety_args: '')
-  workspace = 'Demo.xcworkspace'
-  scheme = 'Demo'
-  configuration = 'Debug'
-  device_host = "platform='iOS Simulator',OS='9.1',name='iPhone 6s'"
+def workspace
+  return 'Demo.xcworkspace'
+end
+
+def configuration
+  return 'Debug'
+end
+
+def targets
+  return [
+    :ios,
+    :osx,
+    :tvos
+  ]
+end
+
+def schemes
+  return {
+    ios: 'Demo',
+    osx: 'MoyaTests-Mac',
+    tvos: 'MoyaTests-tvOS'
+  }
+end
+
+def sdks
+  return {
+    ios: 'iphonesimulator9.1',
+    osx: 'macosx10.11',
+    tvos: 'appletvsimulator9.0'
+  }
+end
+
+def devices
+  return {
+    ios: "OS='9.1',name='iPhone 6s'",
+    osx: "arch='x86_64'",
+    tvos: "OS='9.0',name='Apple TV 1080p'"
+  }
+end
+
+def xcodebuild_in_demo_dir(tasks, platform, xcprety_args: '')
+  sdk = sdks[platform]
+  scheme = schemes[platform]
+  destination = devices[platform]
 
   Dir.chdir('Demo') do
-    sh "set -o pipefail && xcodebuild -workspace '#{workspace}' -scheme '#{scheme}' -configuration '#{configuration}' -sdk iphonesimulator -destination #{device_host} #{tasks} | xcpretty -c #{xcprety_args}"
+    sh "set -o pipefail && xcodebuild -workspace '#{workspace}' -scheme '#{scheme}' -configuration '#{configuration}' -sdk #{sdk} -destination #{destination} #{tasks} | xcpretty -c #{xcprety_args}"
   end
 end
 
 desc 'Build the Demo app.'
 task :build do
-  xcodebuild_in_demo_dir 'build'
+  xcodebuild_in_demo_dir 'build', :ios
 end
 
 desc 'Clean build directory.'
 task :clean do
-  xcodebuild_in_demo_dir 'clean'
+  xcodebuild_in_demo_dir 'clean', :ios
 end
 
 desc 'Build, then run tests.'
 task :test do
-  xcodebuild_in_demo_dir 'build test', xcprety_args: '--test'
+  targets.map { |platform| xcodebuild_in_demo_dir 'build test', platform, xcprety_args: '--test' }
+  sh "killall Simulator"
 end
 
 desc 'Release a version, specified as an argument.'
